+++ projects/egeo/src/lib/st-dropdown-menu/st-dropdown-menu.interface.ts
@@ -20,6 +20,9 @@ export class StDropDownMenuItem {
    textInfo?: string;
    selected?: boolean;
    hasHtml?: boolean;
+   extraIcon?: string;
+   extraIconBubble?: string;
+   extraIconColor?: string;
    [key: string]: any; // To do model more extensible if any other component needs to send more data
 }